backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Session==0.8.0
future==0.18.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgspec==0.18.6
openai==1.16.2
packaging==24.0
pinecone-client==3.2.2
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1